dedupe export transform pipeline into shared helper module
